Update
I hashed out all of the first package loading, since it messed with the figures I wanted to make. Me and Robin think it is because some of the packages are in conflict with eachother, therefore I chose to only load in exactly what I need and skip all the extra stuff.

Best
Philip